Move utils to plugin to provide better way of testing it.
